+++ sql/emaj--4.1.0--devel.sql
@@ -157,6 +157,377 @@ DROP FUNCTION IF EXISTS emaj._build_alter_seq(P_REFLASTVALUE BIGINT,P_REFISCALLE
 ------------------------------------------------------------------
 -- create new or modified functions                             --
 ------------------------------------------------------------------
+CREATE OR REPLACE FUNCTION emaj._check_tables_for_rollbackable_group(p_schema TEXT, p_tables TEXT[], p_arrayFromRegex BOOLEAN,
+                                                                     p_callingFunction TEXT)
+RETURNS TEXT[] LANGUAGE plpgsql AS
+$_check_tables_for_rollbackable_group$
+-- This function filters or verifies that tables are compatible with ROLLBACKABLE groups.
+-- (they must have a PK and no be UNLOGGED or WITH OIDS)
+-- Input: schema, array of tables names, boolean indicating whether the tables list is built from regexp, calling function name
+-- Output: updated tables name array
+  DECLARE
+    v_list                   TEXT;
+    v_array                  TEXT[];
+  BEGIN
+-- Check or discard tables without primary key.
+    SELECT string_agg(quote_ident(relname), ', '), array_agg(relname) INTO v_list, v_array
+      FROM pg_catalog.pg_class t
+           JOIN pg_catalog.pg_namespace ON (pg_namespace.oid = relnamespace)
+      WHERE nspname = p_schema AND t.relname = ANY(p_tables)
+        AND relkind = 'r'
+        AND NOT EXISTS
+              (SELECT 0
+                 FROM pg_catalog.pg_class c
+                      JOIN pg_catalog.pg_namespace ON (pg_namespace.oid = c.relnamespace)
+                      JOIN pg_catalog.pg_constraint ON (connamespace = pg_namespace.oid AND conrelid = c.oid)
+                           WHERE contype = 'p'
+                             AND nspname = p_schema
+                             AND c.relname = t.relname
+              );
+    IF v_list IS NOT NULL THEN
+      IF NOT p_arrayFromRegex THEN
+        RAISE EXCEPTION '%: In schema %, some tables (%) have no PRIMARY KEY.', p_callingFunction, quote_ident(p_schema), v_list;
+      ELSE
+        RAISE WARNING '%: Some tables without PRIMARY KEY (%) are not selected.', p_callingFunction, v_list;
+      -- remove these tables from the tables to process
+      SELECT array_agg(remaining_table) INTO p_tables
+        FROM
+          (  SELECT unnest(p_tables)
+           EXCEPT
+             SELECT unnest(v_array)
+          ) AS t(remaining_table);
+      END IF;
+    END IF;
+-- Check or discard UNLOGGED tables.
+    SELECT string_agg(quote_ident(relname), ', '), array_agg(relname) INTO v_list, v_array
+      FROM pg_catalog.pg_class
+           JOIN pg_catalog.pg_namespace ON (pg_namespace.oid = relnamespace)
+      WHERE nspname = p_schema
+        AND relname = ANY(p_tables)
+        AND relkind = 'r'
+        AND relpersistence = 'u';
+    IF v_list IS NOT NULL THEN
+      IF NOT p_arrayFromRegex THEN
+        RAISE EXCEPTION '%: In schema %, some tables (%) are UNLOGGED tables.', p_callingFunction, quote_ident(p_schema), v_list;
+      ELSE
+        RAISE WARNING '%: Some UNLOGGED tables (%) are not selected.', p_callingFunction, v_list;
+      -- remove these tables from the tables to process
+      SELECT array_agg(remaining_table) INTO p_tables
+        FROM
+          (  SELECT unnest(p_tables)
+           EXCEPT
+             SELECT unnest(v_array)
+          ) AS t(remaining_table);
+      END IF;
+    END IF;
+-- With PG11-, check or discard WITH OIDS tables.
+    IF emaj._pg_version_num() < 120000 THEN
+      SELECT string_agg(quote_ident(relname), ', '), array_agg(relname) INTO v_list, v_array
+        FROM pg_catalog.pg_class
+             JOIN pg_catalog.pg_namespace ON (pg_namespace.oid = relnamespace)
+        WHERE nspname = p_schema
+          AND relname = ANY(p_tables)
+          AND relkind = 'r'
+          AND relhasoids;
+      IF v_list IS NOT NULL THEN
+        IF NOT p_arrayFromRegex THEN
+          RAISE EXCEPTION '%: In schema %, some tables (%) are declared WITH OIDS.', p_callingFunction, quote_ident(p_schema), v_list;
+        ELSE
+          RAISE WARNING '%: Some WITH OIDS tables (%) are not selected.', p_callingFunction, v_list;
+        -- remove these tables from the tables to process
+        SELECT array_agg(remaining_table) INTO p_tables
+          FROM
+           (  SELECT unnest(p_tables)
+            EXCEPT
+              SELECT unnest(v_array)
+           ) AS t(remaining_table);
+        END IF;
+      END IF;
+    END IF;
+--
+    RETURN p_tables;
+  END;
+$_check_tables_for_rollbackable_group$;
+
+CREATE OR REPLACE FUNCTION emaj._assign_tables(p_schema TEXT, p_tables TEXT[], p_group TEXT, p_properties JSONB, p_mark TEXT,
+                                               p_multiTable BOOLEAN, p_arrayFromRegex BOOLEAN)
+RETURNS INTEGER LANGUAGE plpgsql AS
+$_assign_tables$
+-- The function effectively assigns tables into a tables group.
+-- Inputs: schema, array of table names, group name, properties as JSON structure
+--         mark to set for lonnging groups, a boolean indicating whether several tables need to be processed,
+--         a boolean indicating whether the tables array has been built from regex filters
+-- Outputs: number of tables effectively assigned to the tables group
+-- The JSONB p_properties parameter has the following structure '{"priority":..., "log_data_tablespace":..., "log_index_tablespace":...}'
+--   each properties being NULL by default
+  DECLARE
+    v_function               TEXT;
+    v_groupIsRollbackable    BOOLEAN;
+    v_groupIsLogging         BOOLEAN;
+    v_priority               INT;
+    v_logDatTsp              TEXT;
+    v_logIdxTsp              TEXT;
+    v_ignoredTriggers        TEXT[];
+    v_ignoredTrgProfiles     TEXT[];
+    v_list                   TEXT;
+    v_array                  TEXT[];
+    v_timeId                 BIGINT;
+    v_markName               TEXT;
+    v_logSchema              TEXT;
+    v_selectedIgnoredTrgs    TEXT[];
+    v_selectConditions       TEXT;
+    v_eventTriggers          TEXT[];
+    v_oneTable               TEXT;
+    v_nbAssignedTbl          INT = 0;
+  BEGIN
+    v_function = CASE WHEN p_multiTable THEN 'ASSIGN_TABLES' ELSE 'ASSIGN_TABLE' END;
+-- Insert the begin entry into the emaj_hist table.
+    INSERT INTO emaj.emaj_hist (hist_function, hist_event)
+      VALUES (v_function, 'BEGIN');
+-- Check supplied parameters.
+-- Check the group name and if ok, get some properties of the group.
+    PERFORM emaj._check_group_names(p_groupNames := ARRAY[p_group], p_mayBeNull := FALSE, p_lockGroups := TRUE, p_checkList := '');
+    SELECT group_is_rollbackable, group_is_logging INTO v_groupIsRollbackable, v_groupIsLogging
+      FROM emaj.emaj_group
+      WHERE group_name = p_group;
+-- Check the supplied schema exists and is not an E-Maj schema.
+    IF NOT EXISTS
+         (SELECT 0
+            FROM pg_catalog.pg_namespace
+            WHERE nspname = p_schema
+         ) THEN
+      RAISE EXCEPTION '_assign_tables: The schema "%" does not exist.', p_schema;
+    END IF;
+    IF EXISTS
+         (SELECT 0
+            FROM emaj.emaj_schema
+            WHERE sch_name = p_schema
+         ) THEN
+      RAISE EXCEPTION '_assign_tables: The schema "%" is an E-Maj schema.', p_schema;
+    END IF;
+-- Check tables.
+    IF NOT p_arrayFromRegex THEN
+-- From the tables array supplied by the user, remove duplicates values, NULL and empty strings from the supplied table names array.
+      SELECT array_agg(DISTINCT table_name) INTO p_tables
+        FROM unnest(p_tables) AS table_name
+        WHERE table_name IS NOT NULL AND table_name <> '';
+-- Check that application tables exist.
+      WITH tables AS (
+        SELECT unnest(p_tables) AS table_name
+      )
+      SELECT string_agg(quote_ident(table_name), ', ') INTO v_list
+        FROM
+          (SELECT table_name
+             FROM tables
+             WHERE NOT EXISTS
+                     (SELECT 0
+                        FROM pg_catalog.pg_class
+                             JOIN pg_catalog.pg_namespace ON (pg_namespace.oid = relnamespace)
+                        WHERE nspname = p_schema
+                          AND relname = table_name
+                          AND relkind IN ('r','p')
+                     )
+          ) AS t;
+      IF v_list IS NOT NULL THEN
+        RAISE EXCEPTION '_assign_tables: In schema %, some tables (%) do not exist.', quote_ident(p_schema), v_list;
+      END IF;
+    END IF;
+-- Check or discard partitioned application tables (only elementary partitions can be managed by E-Maj).
+    SELECT string_agg(quote_ident(relname), ', '), array_agg(relname) INTO v_list, v_array
+      FROM pg_catalog.pg_class
+           JOIN pg_catalog.pg_namespace ON (pg_namespace.oid = relnamespace)
+      WHERE nspname = p_schema
+        AND relname = ANY(p_tables)
+        AND relkind = 'p';
+    IF v_list IS NOT NULL THEN
+      IF NOT p_arrayFromRegex THEN
+        RAISE EXCEPTION '_assign_tables: In schema %, some tables (%) are partitionned tables (only elementary partitions are supported'
+                        ' by E-Maj).', quote_ident(p_schema), v_list;
+      ELSE
+        RAISE WARNING '_assign_tables: Some partitionned tables (%) are not selected.', v_list;
+        -- remove these tables from the tables to process
+        SELECT array_agg(remaining_table) INTO p_tables
+          FROM
+            (  SELECT unnest(p_tables)
+             EXCEPT
+               SELECT unnest(v_array)
+            ) AS t(remaining_table);
+      END IF;
+    END IF;
+-- Check or discard TEMP tables.
+    SELECT string_agg(quote_ident(relname), ', '), array_agg(relname) INTO v_list, v_array
+      FROM pg_catalog.pg_class
+           JOIN pg_catalog.pg_namespace ON (pg_namespace.oid = relnamespace)
+      WHERE nspname = p_schema
+        AND relname = ANY(p_tables)
+        AND relkind = 'r'
+        AND relpersistence = 't';
+    IF v_list IS NOT NULL THEN
+      IF NOT p_arrayFromRegex THEN
+        RAISE EXCEPTION '_assign_tables: In schema %, some tables (%) are TEMP tables.', quote_ident(p_schema), v_list;
+      ELSE
+        RAISE WARNING '_assign_tables: Some TEMP tables (%) are not selected.', v_list;
+        -- remove these tables from the tables to process
+        SELECT array_agg(remaining_table) INTO p_tables
+          FROM
+           (  SELECT unnest(p_tables)
+            EXCEPT
+              SELECT unnest(v_array)
+           ) AS t(remaining_table);
+      END IF;
+    END IF;
+-- If the group is ROLLBACKABLE, perform additional checks or filters (a PK, not UNLOGGED).
+    IF v_groupIsRollbackable THEN
+      p_tables = emaj._check_tables_for_rollbackable_group(p_schema, p_tables, p_arrayFromRegex, '_assign_tables');
+    END IF;
+-- Check or discard tables already assigned to a group.
+    SELECT string_agg(quote_ident(rel_tblseq), ', '), array_agg(rel_tblseq) INTO v_list, v_array
+      FROM emaj.emaj_relation
+      WHERE rel_schema = p_schema
+        AND rel_tblseq = ANY(p_tables)
+        AND upper_inf(rel_time_range);
+    IF v_list IS NOT NULL THEN
+      IF NOT p_arrayFromRegex THEN
+        RAISE EXCEPTION '_assign_tables: In schema %, some tables (%) already belong to a group.', quote_ident(p_schema), v_list;
+      ELSE
+        RAISE WARNING '_assign_tables: Some tables already belonging to a group (%) are not selected.', v_list;
+        -- remove these tables from the tables to process
+        SELECT array_agg(remaining_table) INTO p_tables
+          FROM
+            (  SELECT unnest(p_tables)
+             EXCEPT
+               SELECT unnest(v_array)
+            ) AS t(remaining_table);
+      END IF;
+    END IF;
+-- Check and extract the tables JSON properties.
+    IF p_properties IS NOT NULL THEN
+      SELECT * INTO v_priority, v_logDatTsp, v_logIdxTsp, v_ignoredTriggers, v_ignoredTrgProfiles
+        FROM emaj._check_json_table_properties(p_properties);
+    END IF;
+-- Check the supplied mark.
+    SELECT emaj._check_new_mark(array[p_group], p_mark) INTO v_markName;
+-- OK,
+    IF p_tables IS NULL OR p_tables = '{}' THEN
+-- When no tables are finaly selected, just warn.
+      RAISE WARNING '_assign_tables: No table to process.';
+    ELSE
+-- Get the time stamp of the operation.
+      SELECT emaj._set_time_stamp('A') INTO v_timeId;
+-- For LOGGING groups, lock all tables to get a stable point.
+      IF v_groupIsLogging THEN
+--   Use a ROW EXCLUSIVE lock mode, preventing for a transaction currently updating data, but not conflicting with simple read access or
+--   vacuum operation.
+        PERFORM emaj._lock_groups(ARRAY[p_group], 'ROW EXCLUSIVE', FALSE);
+--   And set the mark, using the same time identifier.
+        PERFORM emaj._set_mark_groups(ARRAY[p_group], v_markName, FALSE, TRUE, NULL, v_timeId);
+      END IF;
+-- Create new log schemas if needed.
+      v_logSchema = 'emaj_' || p_schema;
+      IF NOT EXISTS
+           (SELECT 0
+              FROM emaj.emaj_schema
+              WHERE sch_name = v_logSchema
+           ) THEN
+-- Check that the schema doesn't already exist.
+        IF EXISTS
+             (SELECT 0
+                FROM pg_catalog.pg_namespace
+                WHERE nspname = v_logSchema
+             ) THEN
+          RAISE EXCEPTION '_assign_tables: The schema "%" should not exist. Drop it manually.',v_logSchema;
+        END IF;
+-- Create the schema.
+        PERFORM emaj._create_log_schema(v_logSchema, CASE WHEN p_multiTable THEN 'ASSIGN_TABLES' ELSE 'ASSIGN_TABLE' END);
+      END IF;
+-- Disable event triggers that protect emaj components and keep in memory these triggers name.
+      SELECT emaj._disable_event_triggers() INTO v_eventTriggers;
+-- Effectively create the log components for each table.
+--   Build the SQL conditions to use in order to build the array of "triggers to ignore at rollback time" for each table.
+      IF v_ignoredTriggers IS NOT NULL OR v_ignoredTrgProfiles IS NOT NULL THEN
+--   Build the condition on trigger names using the ignored_triggers parameters.
+        IF v_ignoredTriggers IS NOT NULL THEN
+          v_selectConditions = 'tgname = ANY (' || quote_literal(v_ignoredTriggers) || ') OR ';
+        ELSE
+          v_selectConditions = '';
+        END IF;
+--   Build the regexp conditions on trigger names using the ignored_triggers_profile parameters.
+        IF v_ignoredTrgProfiles IS NOT NULL THEN
+          SELECT v_selectConditions || string_agg('tgname ~ ' || quote_literal(profile), ' OR ')
+            INTO v_selectConditions
+            FROM unnest(v_ignoredTrgProfiles) AS profile;
+        ELSE
+          v_selectConditions = v_selectConditions || 'FALSE';
+        END IF;
+      END IF;
+-- Process each table.
+      FOREACH v_oneTable IN ARRAY p_tables
+      LOOP
+-- Check that the triggers listed in ignored_triggers property exists for the table.
+        SELECT string_agg(quote_ident(trigger_name), ', ') INTO v_list
+          FROM
+            (  SELECT trigger_name
+                 FROM unnest(v_ignoredTriggers) AS trigger_name
+             EXCEPT
+               SELECT tgname
+                 FROM pg_catalog.pg_trigger
+                      JOIN pg_catalog.pg_class ON (tgrelid = pg_class.oid)
+                      JOIN pg_catalog.pg_namespace ON (relnamespace = pg_namespace.oid)
+                 WHERE nspname = p_schema
+                   AND relname = v_oneTable
+                   AND tgconstraint = 0
+                   AND tgname NOT IN ('emaj_log_trg','emaj_trunc_trg')
+            ) AS t;
+        IF v_list IS NOT NULL THEN
+          RAISE EXCEPTION '_assign_tables: some triggers (%) have not been found in the table %.%.',
+                          v_list, quote_ident(p_schema), quote_ident(v_oneTable);
+        END IF;
+-- Build the array of "triggers to ignore at rollback time".
+        IF v_selectConditions IS NOT NULL THEN
+          EXECUTE format(
+            $$SELECT array_agg(tgname ORDER BY tgname)
+                FROM pg_catalog.pg_trigger
+                     JOIN pg_catalog.pg_class ON (tgrelid = pg_class.oid)
+                     JOIN pg_catalog.pg_namespace ON (relnamespace = pg_namespace.oid)
+                WHERE nspname = %L
+                  AND relname = %L
+                  AND tgconstraint = 0
+                  AND tgname NOT IN ('emaj_log_trg','emaj_trunc_trg')
+                  AND (%s)
+            $$, p_schema, v_oneTable, v_selectConditions)
+            INTO v_selectedIgnoredTrgs;
+        END IF;
+-- Create the table.
+        PERFORM emaj._add_tbl(p_schema, v_oneTable, p_group, v_priority, v_logDatTsp, v_logIdxTsp, v_selectedIgnoredTrgs,
+                              v_groupIsLogging, v_timeId, v_function);
+        v_nbAssignedTbl = v_nbAssignedTbl + 1;
+      END LOOP;
+-- Enable previously disabled event triggers
+      PERFORM emaj._enable_event_triggers(v_eventTriggers);
+-- Adjust the group characteristics.
+      UPDATE emaj.emaj_group
+        SET group_last_alter_time_id = v_timeId,
+            group_nb_table = (
+              SELECT count(*)
+                FROM emaj.emaj_relation
+                WHERE rel_group = group_name
+                  AND upper_inf(rel_time_range)
+                  AND rel_kind = 'r'
+                             )
+        WHERE group_name = p_group;
+-- If the group is logging, check foreign keys with tables outside the groups (otherwise the check will be done at the group start time).
+      IF v_groupIsLogging THEN
+        PERFORM emaj._check_fk_groups(array[p_group]);
+      END IF;
+    END IF;
+-- Insert the end entry into the emaj_hist table.
+    INSERT INTO emaj.emaj_hist (hist_function, hist_event, hist_wording)
+      VALUES (v_function, 'END', v_nbAssignedTbl || ' tables assigned to the group ' || p_group);
+--
+    RETURN v_nbAssignedTbl;
+  END;
+$_assign_tables$;
+
 CREATE OR REPLACE FUNCTION emaj._move_tables(p_schema TEXT, p_tables TEXT[], p_newGroup TEXT, p_mark TEXT, p_multiTable BOOLEAN,
                                              p_arrayFromRegex BOOLEAN)
 RETURNS INTEGER LANGUAGE plpgsql AS
@@ -168,6 +539,7 @@ $_move_tables$
 -- Outputs: number of tables effectively moved to the tables group
   DECLARE
     v_function               TEXT;
+    v_newGroupIsRollbackable BOOLEAN;
     v_newGroupIsLogging      BOOLEAN;
     v_list                   TEXT;
     v_uselessTables          TEXT[];
@@ -175,6 +547,7 @@ $_move_tables$
     v_timeId                 BIGINT;
     v_groups                 TEXT[];
     v_loggingGroups          TEXT[];
+    v_nbAuditOnlyGroups      INT;
     v_groupName              TEXT;
     v_groupIsLogging         BOOLEAN;
     v_oneTable               TEXT;
@@ -186,7 +559,7 @@ $_move_tables$
       VALUES (v_function, 'BEGIN');
 -- Check the group name and if ok, get some properties of the group.
     PERFORM emaj._check_group_names(p_groupNames := ARRAY[p_newGroup], p_mayBeNull := FALSE, p_lockGroups := TRUE, p_checkList := '');
-    SELECT group_is_logging INTO v_newGroupIsLogging
+    SELECT group_is_rollbackable, group_is_logging INTO v_newGroupIsRollbackable, v_newGroupIsLogging
       FROM emaj.emaj_group
       WHERE group_name = p_newGroup;
 -- Check the tables list.
@@ -232,11 +605,11 @@ $_move_tables$
           WHERE tbl <> ALL(v_uselessTables);
       END IF;
     END IF;
--- Get the lists of groups and logging groups holding these tables, if any.
--- It locks the tables groups so that no other operation simultaneously occurs these groups
+-- Get the lists of groups and logging groups holding these tables, if any, and count the number of AUDIT_ONLY groups.
+-- It locks the target and source tables groups so that no other operation simultaneously occurs these groups
 -- (the CTE is needed for the FOR UPDATE clause not allowed when aggregate functions).
     WITH tables_group AS (
-      SELECT group_name, group_is_logging FROM emaj.emaj_group
+      SELECT group_name, group_is_logging, group_is_rollbackable FROM emaj.emaj_group
         WHERE group_name = p_newGroup OR
               group_name IN
                (SELECT DISTINCT rel_group FROM emaj.emaj_relation
@@ -246,9 +619,15 @@ $_move_tables$
         FOR UPDATE OF emaj_group
       )
     SELECT array_agg(group_name ORDER BY group_name),
-           array_agg(group_name ORDER BY group_name) FILTER (WHERE group_is_logging)
-      INTO v_groups, v_loggingGroups
+           array_agg(group_name ORDER BY group_name) FILTER (WHERE group_is_logging),
+           count(group_name) FILTER (WHERE NOT group_is_rollbackable AND group_name <> p_newGroup)
+      INTO v_groups, v_loggingGroups, v_nbAuditOnlyGroups
       FROM tables_group;
+-- If at least 1 source tables group is of type AUDIT_ONLY and the target tables group is ROLLBACKABLE, add some checks on tables.
+-- They may be incompatible with ROLLBACKABLE groups.
+    IF v_nbAuditOnlyGroups > 0 AND v_newGroupIsRollbackable THEN
+      p_tables = emaj._check_tables_for_rollbackable_group(p_schema, p_tables, p_arrayFromRegex, '_move_tables');
+    END IF;
 -- Check the supplied mark.
     SELECT emaj._check_new_mark(v_loggingGroups, p_mark) INTO v_markName;
 -- OK,